zsh: conform to XDG standard
Why I tried to use `trash-cli` installed by `pipx`,
I find the script crashed because of a space in the Shebang.
Unfortunately there's no techniques to escape a space in Shebang.
So I must configure `PIPX_HOME` for `pipx`.
That's why I begin to notice the importance of XDG standard.

I plan to put all XDG-supported configurations in `.zshenv`,
as it will be also sourced in a non-interactive shell.