nvim: remove autocmds that restart yabai / reload skhd config
These autocmds aim to automatically reload yabai and skhd when the configurations changes.

However, I find if the reload process is executed by neovim, the mouse behavior becomes somehow strange.
For instance, when I move my mouse to the border of the window, the mouse type doesn't change -
the range in which the mouse type can change is notably shrunk.
I'm not sure whether it's a bug of yabai or neovim. So I remove these two autocmds temporarily.