zsh(plugin): fix nnn plugin
Before quitting, nnn will write a `cd` command to the file `$NNN_TMPFILE`.
A shell can execute the command to `cd` to the final directory.

However, if the dir doesn't change at all, by default the shell will perform the `cd` anyway.
This behavior may push duplicated entries in the directory stack of some shell (like zsh).

To address the problem w/o touching the source code of nnn, we need to parse the `cd` command,
find out the final directory, and check if the dir has changed before trying to `cd`.

The most annoying part is to handle the escaped `'` since nnn has "kindly" performed shell escape.
(The contributor admitted that it's a design mistake, but kept for backward compatibility.)

The workaround seems to be dirty, but it conforms to the design goal of nnn: feature-restricted.