zsh(fix): Emacs vterm issue
In Emacs vterm, there is no way to distinguish "ctrl-j" and "<RET>". This is
surprising, and I do not know the underlying reason currently. So I disable the
key mapping of "accept-and-hold" when the Zsh is opened inside Emacs vterm.

Meanwhile, the color of zsh-autosuggestion is strange. Add underline to it.

Ref: https://emacs-china.org/t/vterm-zsh/20497/8