fix(vterm)!: add module emacs-libvterm to manage shell integration
The new module emacs-libvterm replaces the old way of managing vterm shell
integration (as a Zsh plugin).

The shortcoming of using zsh plugin system: it's difficult to specify the order
of loading plugins, and the Nix logic is coupled with the Zsh init file's.

Vterm shell integration must be loaded after starship is loaded, so it can
hooked into the shell prompt.