zsh: use ^O instead of ^J for menuselect key binding
In modern sane tty, ICRNL is set by default so the control character CR (^M and
the return key) will be converted to NL (^J) instead. It's hard to distinguish
CR and NL most of the time. I've tried `stty -icrnl` but it cannot be recognized
correctly by the vterm function `vterm--get-icrnl`, and it may trigger further
problems :(

Therefore I rebind `accept-and-infer-next-history` to ^O in menuselect.