font: move font packages from nix-darwin side to home-manager side
I just found managing fonts with home-manager is easy: just append them in
`home.packages`, and the activation script will copy all font files to
~/Library/Fonts/HomeManager.

Managing fonts with nix-darwin is also a good idea, but it uses a more
complicated script and the hash in Nix store is kept, making the path really
ugly. So I switch to home-manager for font managing.