git: avoid overriding drv in case of rebuilding git
`withSsh` is a misleading option name — turning this option off will NOT disable
ssh feature for the git release. In fact, if this option is true, git will be
built with ssh path hardcoded. Generally, we have `ssh` in our PATH env, so
there is no need to hardcode such its path.

Setting this option to true triggers rebuilding of git which really takes a lot
of time. So I remove the overriding to benefit from the cache.